requirements.txt
actionlib==1.13.2
Adafruit-Blinka==6.12.0
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
angles==1.9.13
asn1crypto==0.24.0
bcrypt==3.1.6
catkin==0.8.10
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
coverage==4.5.2
cryptography==2.6.1
defusedxml==0.5.0
docutils==0.14
empy==3.3.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gpiozero==1.5.1
idna==2.6
message-filters==1.15.11
mock==2.0.0
netifaces==0.10.4
nose==1.3.7
numpy==1.16.2
olefile==0.46
paho-mqtt==1.5.1
paramiko==2.4.2
pbr==4.2.0
Pillow==5.4.1
pyasn1==0.4.2
pycryptodomex==3.6.1
pyftdi==0.53.2
Pygments==2.3.1
PyNaCl==1.3.0
pyparsing==2.2.0
pyserial==3.5
python-apt==1.8.4.3
python-dateutil==2.7.3
python-gnupg==0.4.4
pyusb==1.1.0
PyYAML==3.13
raspberrypi-esp32-i2c==0.0.6
requests==2.21.0
roman==2.0.0
rosbag==1.15.11
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.11
roslaunch==1.15.11
roslib==1.15.7
roslz4==1.15.11
rosmake==1.15.7
rosmaster==1.15.11
rosmsg==1.15.11
rosnode==1.15.11
rosparam==1.15.11
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.15.11
rosserial-arduino==0.9.2
rosserial-client==0.9.2
rosserial-python==0.9.2
rosserial-vex-cortex==0.9.2
rosserial-vex-v5==0.9.2
rosserial-xbee==0.9.2
rosservice==1.15.11
rostest==1.15.11
rostopic==1.15.11
rosunit==1.15.7
roswtf==1.15.11
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
sensor-msgs==1.13.1
six==1.12.0
smbus2==0.4.1
spidev==3.4
ssh-import-id==5.7
sysv-ipc==1.1.0
tf==1.13.2
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.11
urllib3==1.24.1
